act 6 step 1-2-3

+++ app/Http/Controllers/ProductController.php
@@ -62,6 +62,8 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        $this->authorize('manage-product', $product);
+
         return view('products.edit', compact('product'));
     }
 
@@ -70,6 +72,8 @@ class ProductController extends Controller
      */
     public function update(Request $request,  Product $product)
     {
+        $this->authorize('manage-product', $product);
+
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
@@ -91,6 +95,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        $this->authorize('manage-product', $product);
+        
         $product->delete();
 
         return redirect()
